// apps/autoops-web/generated/api.ts
export type UpdateSystemSettingsBodyGeneral = {
  siteName?: string;
  maintenanceMode?: boolean;
};

export type UpdateSystemSettingsBodyNotifications = {
  certExpiryReminder?: boolean;
  healthCheckAlert?: boolean;
};

export type UpdateSystemSettingsBodySystem = {
  autoCleanupLogs?: boolean;
  appVersion?: string;
  nodeVersion?: string;
  database?: string;
  databaseVersion?: string;
  redis?: string;
  redisVersion?: string;
};

export type UpdateSystemSettingsBody = {
  general?: UpdateSystemSettingsBodyGeneral;
  notifications?: UpdateSystemSettingsBodyNotifications;
  system?: UpdateSystemSettingsBodySystem;
};